+++ python-m1/pycharm/ex/ex010.py
@@ -1,4 +1,4 @@
 print(f'Conversor Reais para Dollar')
-brl = int(input('Quantos reais voce tem?: '))
-usd = int(brl) / 5.76
+brl = float(input('Quantos reais voce tem?: '))
+usd = brl / 5.76
 print(f'Voce pode comprar {usd:.2f} dolares.')
